requirements.txt
Django>=5.2,<6
djangorestframework>=3.15,<3.16
firebase-admin>=6.2,<7
pdfplumber>=0.11,<0.12
PyPDF2>=3,<4
pytesseract>=0.3,<0.4
Pillow>=10,<11
reportlab>=4,<5
google-generativeai>=0.7,<0.8
python-json-logger==2.0.7
sentry-sdk[django]==2.0.1
django-prometheus==2.3.1
python-dotenv>=1.0,<2
ruff>=0.6,<0.7
colorlog>=6.9,<7
django-cors-headers>=4.4,<5
django-filter>=24.2,<25
psycopg2-binary>=2.9,<3
resend>=2.0,<3
inflection==0.5.1